Testing updates
[mxpf]
- Now resets mxPatchFile to nil when FinalizeMXPF is called.
- New boolean variable: mxDisallowSaving.  Stops MXPF from saving logs
when finalized even if the user constant is set to true.  For use when
running tests to avoid producing a billion log files.
- New boolean variable: mxHideErrorPopups.  Stops ShowMessage calls when
errors occur.  For use whne running tests.
- New boolean variable: mxDisallowNewFile.  Stops the creation of new
files by the PatchFileByAuthor/PathFileByName methods when a matching
file isn't found.  I'm mainly using this for when running tests, but it
could be useful in other situations too.
- LoadRecords will no longer attempt to access the filename of
mxPatchFile when it isn't assigned.

[jvTest]
- New function ExpectEqual: takes two variant values and a test string.
If the variants are not equal, it will attempt to throw a meaningful
exception stating their values if they are of a supported type (string,
double, integer).

[mxpfTests]
- Now using clean and consistent formatting conventions to make the code
more readable.
- Added tests for patch file selection, file selection, and LoadRecords
- Fleshed out Test plan.txt some more